+++ src/api/index.ts
@@ -9,6 +9,7 @@ export * from './operators';
 export * from './rotations';
 export * from './sizes';
 export * from './snapshot';
+export * from './extraction';
 export * from './PDFPageOptions';
 export * from './PDFDocumentOptions';
 export * from './StandardFonts';